Filtragem dos anagramas funciona, mas não é suficiente.
